+++ src/xrCDB/Frustum.cpp
@@ -82,7 +82,7 @@ u32 frustum_aabb_remap[8][6] =
 EFC_Visible CFrustum::testSphere(Fvector& c, float r, u32& test_mask) const
 {
     u32 bit = 1;
-    for (int i = 0; i < p_count; i++, bit <<= 1)
+    for (size_t i = 0; i < p_count; i++, bit <<= 1)
     {
         if (test_mask & bit)
         {
@@ -112,7 +112,7 @@ EFC_Visible CFrustum::testAABB(const float* mM, u32& test_mask) const
     // go for trivial rejection or acceptance using "faster overlap test"
     u32 bit = 1;
 
-    for (int i = 0; i < p_count; i++, bit <<= 1)
+    for (size_t i = 0; i < p_count; i++, bit <<= 1)
     {
         if (test_mask & bit)
         {
@@ -132,7 +132,7 @@ EFC_Visible CFrustum::testAABB(const float* mM, u32& test_mask) const
 EFC_Visible CFrustum::testSAABB(Fvector& c, float r, const float* mM, u32& test_mask) const
 {
     u32 bit = 1;
-    for (int i = 0; i < p_count; i++, bit <<= 1)
+    for (size_t i = 0; i < p_count; i++, bit <<= 1)
     {
         if (test_mask & bit)
         {
@@ -163,7 +163,7 @@ EFC_Visible CFrustum::testSAABB(Fvector& c, float r, const float* mM, u32& test_
 bool CFrustum::testPolyInside_dirty(Fvector* p, size_t count) const
 {
     Fvector* e = p + count;
-    for (int i = 0; i < p_count; i++)
+    for (size_t i = 0; i < p_count; i++)
     {
         const fplane& P = planes[i];
         for (Fvector* I = p; I != e; I++)
@@ -334,7 +334,7 @@ sPoly* CFrustum::ClipPoly(sPoly& S, sPoly& D) const
 {
     sPoly* src = &D;
     sPoly* dest = &S;
-    for (int i = 0; i < p_count; i++)
+    for (size_t i = 0; i < p_count; i++)
     {
         // cache plane and swap lists
         const fplane& P = planes[i];
@@ -480,7 +480,7 @@ void CFrustum::CreateFromMatrix(Fmatrix& M, u32 mask)
         p_count++;
     }
 
-    for (int i = 0; i < p_count; i++)
+    for (size_t i = 0; i < p_count; i++)
     {
         float denom = 1.0f / planes[i].n.magnitude(); // Get magnitude of Vector
         planes[i].n.x *= denom;
